requirements.txt
flask>=2.0.1
flask-cors>=3.0.10
python-dotenv>=0.19.0
openai>=1.3.0
google-generativeai>=0.1.0
Flask-Limiter>=3.5.0
redis>=4.5.1 